Move methods and tests to COTMigrationUtil
